+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased][unreleased]
 
+### Changed
+
+- Added Manage projects and cases
+- 
 ## [Release-44][release-44] (production-2025-08-06.6211)
 
 ### Changed